chore: abstract test constants out to __mocks__
